Feature/16272 adjust retry (#8)
Adjust retry mechanism
