+++ omnia_timeseries/http_client.py
@@ -3,40 +3,12 @@ from azure.identity._internal.msal_credentials import MsalCredential
 import requests
 import logging
 from omnia_timeseries.helpers import retry
-from requests.models import Response
-import json
+from omnia_timeseries.models import TimeseriesRequestFailedException
 from importlib import metadata
 import sys
 
 logger = logging.getLogger(__name__)
 version = metadata.version("omnia-timeseries-api")
-
-
-class TimeseriesRequestFailedException(Exception):
-    def __init__(self, response: Response) -> None:
-        error = json.loads(response.text)
-        self._status_code = response.status_code
-        self._reason = response.reason
-        self._message = error["message"]
-        self._trace_id = error["traceId"]
-        super().__init__(
-            f"Status code: {self._status_code}, Reason: {self._reason}, Message: {self._message}, Trace ID: {self._trace_id}")
-
-    @property
-    def status_code(self) -> int:
-        return self._status_code
-
-    @property
-    def reason(self) -> str:
-        return self._reason
-
-    @property
-    def message(self) -> str:
-        return self._message
-
-    @property
-    def trace_id(self) -> str:
-        return self._trace_id
 
 
 @retry(logger=logger)
