+++ omnia_timeseries/models.py
@@ -1,29 +1,38 @@
 from typing import List, Optional, TypedDict, Dict, Any, Literal
+from requests.models import Response
+import json
+
 
 class MessageModel(TypedDict):
     statusCode: int
     message: str
     traceId: str
 
+
 class DatapointModel(TypedDict):
     time: str
     value: float
     status: int
 
+
 class DatapointsPostRequestModel(TypedDict):
     datapoints: List[DatapointModel]
+
 
 class DatapointsItemModel(TypedDict):
     id: str
     datapoints: List[DatapointModel]
 
+
 class DatapointsItemsModel(TypedDict):
     items: List[DatapointsItemModel]
+
 
 class GetDatapointsResponseModel(TypedDict):
     data: DatapointsItemsModel
     count: Optional[int]
     continuationToken: Optional[str]
+
 
 class GetMultipleDatapointsRequestItem(TypedDict, total=False):
     id: str
@@ -33,9 +42,11 @@ class GetMultipleDatapointsRequestItem(TypedDict, total=False):
     statusFilter: Optional[List[int]]
     includeOutsidePoints: Optional[bool]
     fill: Optional[str]
-    aggregateFunction: Optional[List[Literal['avg', 'min', 'max', 'sum', 'stddev', 'count', 'first', 'last']]]
+    aggregateFunction: Optional[List[Literal['avg', 'min',
+                                             'max', 'sum', 'stddev', 'count', 'first', 'last']]]
     processingInterval: Optional[str]
     continuationToken: Optional[str]
+
 
 class AggregateModel(TypedDict, total=False):
     time: str
@@ -50,17 +61,21 @@ class AggregateModel(TypedDict, total=False):
     value: Optional[float]
     status: Optional[int]
 
+
 class AggregateItemModel(TypedDict):
     id: str
     datapoints: List[AggregateModel]
 
+
 class AggregateItemsModel(TypedDict):
     items: List[AggregateItemModel]
+
 
 class GetAggregatesResponseModel(TypedDict):
     data: AggregateItemsModel
     count: Optional[int]
     continuationToken: Optional[str]
+
 
 class TimeseriesModel(TypedDict):
     id: str
@@ -76,13 +91,16 @@ class TimeseriesModel(TypedDict):
     source: str
     metadata: Optional[Dict[str, Any]]
 
+
 class TimeseriesItemsModel(TypedDict):
     items: List[TimeseriesModel]
+
 
 class GetTimeseriesResponseModel(TypedDict):
     data: TimeseriesItemsModel
     count: Optional[int]
     continuationToken: Optional[str]
+
 
 class TimeseriesRequestItem(TypedDict, total=False):
     name: str
@@ -94,6 +112,7 @@ class TimeseriesRequestItem(TypedDict, total=False):
     externalId: Optional[str]
     metadata: Optional[Dict[str, Any]]
 
+
 class TimeseriesPatchRequestItem(TypedDict, total=False):
     name: Optional[str]
     description: Optional[str]
@@ -104,38 +123,75 @@ class TimeseriesPatchRequestItem(TypedDict, total=False):
     externalId: Optional[str]
     metadata: Optional[Dict[str, Any]]
 
+
 class FacilityModel(TypedDict):
     facility: str
     tsdbExportEnabled: float
     dlExportEnabled: float
-    totalImsSubscriptions:float 
+    totalImsSubscriptions: float
+
 
 class FacilityItemsModel(TypedDict):
     items: List[FacilityModel]
 
+
 class FacilityDataModel(TypedDict):
     data: FacilityItemsModel
+
 
 class HistoryItem(TypedDict):
     changedBy: str
     changedTime: str
     changes: Dict[str, dict]
 
+
 class HistoryItemsModel(TypedDict):
     items: List[HistoryItem]
+
 
 class GetHistoryResponseModel(TypedDict):
     data: HistoryItemsModel
 
+
 class StreamSubscriptionRequestModel(TypedDict):
     id: str
+
 
 class StreamSubscriptionModel(TypedDict):
     id: str
     customerId: str
 
+
 class StreamSubscriptionItemsModel(TypedDict):
     items: List[StreamSubscriptionModel]
 
+
 class StreamSubscriptionDataModel(TypedDict):
     data: StreamSubscriptionItemsModel
+
+
+class TimeseriesRequestFailedException(Exception):
+    def __init__(self, response: Response) -> None:
+        error = json.loads(response.text)
+        self._status_code = response.status_code
+        self._reason = response.reason
+        self._message = error["message"]
+        self._trace_id = error["traceId"]
+        super().__init__(
+            f"Status code: {self._status_code}, Reason: {self._reason}, Message: {self._message},  Trace ID: {self._trace_id}")
+
+    @property
+    def status_code(self) -> int:
+        return self._status_code
+
+    @property
+    def reason(self) -> str:
+        return self._reason
+
+    @property
+    def message(self) -> str:
+        return self._message
+
+    @property
+    def trace_id(self) -> str:
+        return self._trace_id
