+++ tests/test_api.py
@@ -3,10 +3,10 @@ import requests_mock
 import pytest
 from azure.identity._internal.msal_credentials import MsalCredential
 from azure.core.credentials import AccessToken
-
 from requests.models import Response
 from omnia_timeseries import TimeseriesAPI, TimeseriesEnvironment
 from omnia_timeseries.http_client import TimeseriesRequestFailedException
+
 
 class DummyCredentials(MsalCredential):
 
@@ -17,15 +17,26 @@ class DummyCredentials(MsalCredential):
         return None
 
 
+@pytest.fixture
+def api():
+    env = TimeseriesEnvironment("dummy", "https://test")
+    api = TimeseriesAPI(DummyCredentials("dummy"), env)
+    return api
 
-class ApiTestCase(unittest.TestCase):
 
-    def test_retry_on_failure(self):
-        env = TimeseriesEnvironment("dummy", "https://test")
-        api = TimeseriesAPI(DummyCredentials("dummy"), env)
-        with requests_mock.Mocker() as m:
-            m.register_uri("GET", "https://test/1234/data/latest", status_code=503, text="""{"message": "Service is unavailable", "traceId": "1"}""")
-            with pytest.raises(TimeseriesRequestFailedException):
-                response = api.get_latest_datapoint("1234")
-        assert m.call_count == 3, "Unexpected number of retries"
-            
+def test_retry_on_failure(api):
+    with requests_mock.Mocker() as m:
+        m.register_uri("GET", "https://test/1234/data/latest", status_code=503,  # 503 is retryable
+                       text="""{"message": "Service is unavailable", "traceId": "1"}""")
+        with pytest.raises(TimeseriesRequestFailedException):
+            api.get_latest_datapoint("1234")
+    assert m.call_count == 3, "Unexpected number of retries"
+
+
+def test_skip_retry_when_not_retryable_status_code(api):
+    with requests_mock.Mocker() as m:
+        m.register_uri("GET", "https://test/1234/data/latest", status_code=403,  # 403 is not retryable
+                       text="""{"message": "Service is unavailable", "traceId": "1"}""")
+        with pytest.raises(TimeseriesRequestFailedException):
+            api.get_latest_datapoint("1234")
+    assert m.call_count == 1, "Unexpected number of retries"
